Source/collaboration_2/requirements.txt
Flask==1.0
psycopg2
sqlalchemy
sqlalchemy_utils
flask-sqlalchemy
flask-restful
flask-cors
requests
pandas==0.22.0
Image
pyzmq
locustio
xlsxwriter
Pillow
flask_cache
validators
numpy
openpyxl
flask_httpauth
itsdangerous
flask-migrate
flask-script
whoosh
xlrd
diff-match-patch
Elasticsearch
flask_apscheduler